+++ runner2.py
@@ -14,7 +14,7 @@ def maxCeldasPorSeg(res):
     for linea in res.stderr.splitlines():
         if "celdas incendiadas por microsegundo" in linea:
             partes = linea.strip().split()
-            sumaCeldas.append(double(partes[-1]))
+            sumaCeldas.append(float(partes[-1]))
     return max(sumaCeldas)
 
 def main():
